Add .vscode/settings.json. (#263)
This file is gitignored, so it will provide default settings that can
be further customized without showing up as pending changes.

It enables `editor.formatOnSave` to help with clang-format enforcement.

It sets `files.eol` to CRLF and enables `files.insertFinalNewline`,
`files.trimFinalNewlines`, and `files.trimTrailingWhitespace` to help
with whitespace enforcement.

It uses `files.associations` to mark our extensionless headers as C++.

It excludes the `stl/msbuild` subdirectory (which isn't used by the
GitHub repo, and is doomed soon). It also excludes the `vcpkg`
submodule, to prevent its files from showing up in Quick Open (Ctrl+P).