Fix span CTAD from built-in arrays.
This is a defect in the Standardese, for which I've submitted an LWG
issue with an extensive writeup. Basically, the constrained constructor
out-competes the non-constrained deduction-guide, resulting in the
dynamic_extent default argument being used. Adding type_identity_t
prevents the constructor from interfering with CTAD (which is why the
std::array constructors aren't affected; they involve value_type which
is remove_cv_t<ElementType>).

This is specific to concepts; there's Standardese that prefers
"more constrained" during overload resolution. The non-concepts
implementation of this constructor doesn't need to be changed.