Add the tr1 test suite. (#485)
* Add the tr1 test suite.

Works towards #144. Note that this is just the source code; the
infrastructure to automatically build and run the tests is coming soon.

Update enforce-clang-format.cmd as parallelize is recursive. (I verified
that behavior by running parallelize directly.) We don't need to mention
stl/inc's subdirectories. We can mention all of tools because
clang-format detects file extensions and the other files in tools are
fine. (In contrast, excluding stl/msbuild saves time.) Finally, add
tests/tr1 (which is why we depend on recursive behavior, due to its many
subdirectories). Also, update the git status and git diff commands to
print differences in the tests directory too.

* Add "accepted extensions" to parallelize.