Fix sign-conversion warnings in the test.
When constants like `2` are directly passed to functions taking
`size_t`, the compiler can see that the value is being preserved,
so it doesn't warn. However, when `2` is perfectly forwarded,
only the type `int` is forwarded - the constant nature isn't part
of the type system. So, the compiler will warn if it sees a forwarded
`int` converted to `size_t`.

The fix is to use `2u` etc. in the source code. `unsigned int`
being converted to `size_t` is known to be value-preserving
(for 64-bit `size_t`, it's a widening).