Fix sign-conversion warning, use C++ casts.
First, this was using functional-style C-semantics casts to convert
from unsigned to signed. We conventionally avoid all C-semantics casts
regardless of style (except `(void)`).

Second, this was implicitly converting from signed back to unsigned,
emitting warnings for which we attempt to be clean (even though
they're off-by-default).

Finally, this had unnecessary parentheses in the conditional operator;
it's a judgement call, but stylistically we tend to avoid parentheses
given these operators.