Adjust comments and class/struct.
In new code, we generally prefer to "detach" introductory comments
from braces, especially when that reduces wrapping for the comment
and the function parameters.

Conversely, we generally prefer to attach TRANSITION comments to the
affected lines, so we can easily see what needs to be adjusted
when the mentioned bug is fixed.

`_Uninitialized_rev_destroying_backout_al` is a class that uses
private access control. Therefore, its comment should refer to it as a
class. Additionally, we generally avoid relying on classes defaulting
to private access control; this is one of the relatively few places
where we prefer to repeat the default.