`_Allocate_flexible_array` refactoring, part 1.
This carefully performs multiplication and addition overflow checks,
and it avoids the use of `new[]` with aligned_storage_t that was
emitting buffer overrun warnings. The new strategy is to give
`_Ref_count_unbounded_array` storage for one element, and simply
allocate additional space (via plain ::operator new) for additional
elements if necessary. We need a bit of extra logic for alignment,
performed with bitwise operations here (to avoid dividing and
multiplying, even by a constant power of 2). (The concern is that if
the refcount control block has an alignment of 8, and the elements have
a size like 1, then the overall space requested might be an odd number
like 1729, which might cause ::operator new to think that it doesn't
need to return 8-aligned storage. By rounding up to the maximum
alignment of the control block and the element, we avoid that issue.)

This doesn't yet use an RAII guard for ::operator delete, but I plan
to do so in the near future.

This uses a non-_Ugly typedef element_type in internal code. This is
slightly non-conventional, but safe (because it's a Standard typedef
in shared_ptr, it can't be macroized). There's some existing precedent
for this, and I wanted to reinforce that "this is the usual
element_type definition that you've seen elsewhere".

Part 2 will extend this to the allocator control blocks.