`_Allocate_flexible_array` refactoring, part 2.
This extends the refactoring to allocate_shared. It uses a "hybrid"
approach, where the control block contains a union of one element
(treated as a flexible array), giving the control block proper
alignment, but storage for the control block is allocated as a
sequence of `aligned_storage_t<_Align, _Align>` elements.

This should resolve all overflow issues. Overflow checking is
controllable because we don't need it during cleanup.

Add comments citing N4849 [class.dtor]/7, explaining why these
destructors must be defined with `{}` instead of `= default;`.