Silence warning in constant-evaluated branch of _Find_unchecked1
When doing memchr optimizations in find we fallback to the
standard find algorithm when the values are eligible for memchr
optimization but we are constant evaluated. This can result in
emitting sign-mismatch warnings which are pointless since we already
do bounds checking of the search element.

Just cast the search element to the type of the container's values.