Silence warning in execution
One of our tests instantiates a _Static_partition_team with the type of _Count
as a short. We store the result of a short-short division and a short-short
mod operation into a short which results in a narrowing warning from int to short.
This is due to short to int promotion for arithmetic binary operations.