Always use _Bit_cast.
Copy bit_cast in <bit> to _Bit_cast in <xutility>, superseding _Bit_cast
in <charconv>. Unlike C++20 bit_cast, _Bit_cast is always available.
(<xutility> is a common ancestor of <complex> and <charconv>, and
includes the necessary machinery for this definition.) I verified that
MSVC, Clang, and IntelliSense all accept `__builtin_bit_cast` in 14/17
mode. For CUDA, I provide the memcpy fallback (which isn't constexpr,
of course).

Finally, _Isinf should use _Bit_cast; I'm marking that with a comment
(and will file a GitHub issue in the future) instead of extending this
PR indefinitely.