Remove _For_each_tuple_element in scoped_lock
scoped_lock::~scoped_lock uses _For_each_tuple_element which is not
needed in C++17 because fold expressions can be used. This PR removes
_For_each_tuple_element from tuple and changes the dtor in scoped_lock
to use fold expression with std::apply.

To minimize the changes in test (VSO_0000000_instantiate_containers)
using _For_each_tuple_element, we copied this function to the test.