Cleanup: Eliminate macros, use alias templates.
This eliminates the _GENERIC_MATHC0X and _GENERIC_MATHC1X macros in
favor of directly stamping out the functions. (Over the years, we've
moved away from the practice of using macros to stamp out highly
repetitive code. The only remaining scenario where I believe that
macros are valuable is when we need to stamp out something for every
possible calling convention etc. as that's incredibly verbose and
platform-dependent.)

Additionally, this changes _Promote_to_float to _Upgrade_to_double.
This is better-named in two different ways ("promote" is a Word Of
Power that isn't exactly applicable here, and the replacement type is
double). It's also less verbose and higher throughput (the former is
obvious; the latter is because struct templates are more expensive for
compilers to instantiate, as they have to do a lot of work when they
generate a class definition).

Avoiding macros allows us to perform a couple of micro-cleanups: we can
directly return 0 (no need to static_cast to floating type) and we
don't have to parenthesize `_Left * _Left` to make clang-format happy.