Change `if (_Code != error_code())` to `if (_Code)`.
This is more efficient (because it avoids calling `system_category()`)
and consistent with Standard `<filesystem>`. There's a slight semantic
difference, which I believe is either irrelevant here or is an
improvement: `error_code` equality considers both `value()` and
`category()`, so an `error_code` with `value() == 0` but
`generic_category()` would compare non-equal with `error_code()`, yet
it would test as `false`. I believe that this is irrelevant because the
`<experimental/filesystem>` machinery will never construct such an
`error_code`.