Remove using-directive in `<regex>`.
This was the only occurrence of a "convenient" using-directive in the
entire STL's product code. (The others are doing necessary work,
or working around compiler bugs.)

While this was function-local and didn't create correctness issues,
I felt that it was inconsistent with our usual conventions, and wasn't
even saving significant verbosity. (Long ago, this was an enormously
nested conditional expression, where the using-directive indeed helped.
That was changed by Microsoft-internal MSVC-PR-182277, to the current
chain of if-statements.)