<variant>: silence useless warning
We don't want narrowing (or any other, for that matter) conversion warnings in the machinery that determines which variant alternative to activate given an initializer. The actual initialization code should certainly emit these warnings, but emitting them from the metaprogramming machinery can produce in warnings early in overload resolution while determining conversion sequences for overloads that might not be selected.

Fixes DevCom-1031281.

Drive-by: convert `_Is_in_place_index` into a variable template.