Partially fix handling of `-0` in <complex>
Partially address #67.

Use `_Abs` and `_Copysign` instead of branching on `x < 0`, correctly
treating negative zero as a negative number. This matters when branch
cuts of a complex function lies on the real or the imaginary axis, where
the results for positive zero and negative zero are different.

The following cases are not fixed in this commit:

* `pow` when the first argument is (negative - 0i).

* (+/-∞ - 0i) and (-0 +/- ∞i).