Copy-paste fix! Change 64 to 32.
The 1, 4, and 8-byte reverse_copy implementations all start by testing
for 32 bytes of AVX2 work. This is exactly half of what the 1/2/4/8-byte
reverse implementations test for; I believe they need 64 bytes because
they swap the front and the back of the range (so, 32 bytes on each
side).

Uniquely, 2-byte reverse_copy said 64. I believe that this is a
copy-paste error, and that it should say 32 like all other reverse_copy
implementations.

This couldn't have been caught by testing, because the effect was to
disable the AVX2 optimization for 2-byte elements, where the source
range was between [32, 64) bytes, resulting in reduced performance only.