Fix queue/stack friendship.
queue<T, NonSpaceshipContainer> might be instantiated for a
non-three_way_comparable NonSpaceshipContainer. In that case, it
shouldn't attempt to grant friendship to the spaceship operator,
because that would attempt to form
compare_three_way_result_t<NonSpaceshipContainer>.

As far as I can tell, there's no way in the Core Language for a queue
to grant friendship to only one specialization of its non-member
spaceship operator while respecting the constraint. (See
WG21-N4861 [temp.friend]/9.) Instead, we need to template the
friendship declaration, such that queue grants friendship to all of its
spaceship operators. This is safe, just verbose.