Further improve `_Vbase_compare_three_way` codegen.
Testing `_Differing_bits == 0` might seem like it's asking the compiler
to do more work (always XORing). But it allows the compiler to see that
the input to `_Countr_zero` is nonzero, and that allows it to optimize
away the runtime ISA check, unconditionally emitting a `tzcnt`
instruction.

Then, at the end of the function, we know something that the compiler
doesn't realize: `_Mask` selects a differing bit (the least significant
one), so `(_Left & _Mask)` and `(_Right & _Mask)` are zero/nonzero
or vice versa. Therefore, we only need to inspect one side!

I've verified that each of these changes improves the codegen.

Co-authored-by: Casey Carter <Casey@Carter.net>