Fix a subtle conformance issue.
It's technically nonconformant to explicitly specialize a non-inline
variable template (and _INLINE_VAR is inline for C++17 and above). MSVC
accepts this, but it causes problems for Clang.
tests/std/tests/Dev09_172666_tr1_tuple_odr detected this in the
MSVC-internal repo, but not on GitHub due to an oversight that we just
discovered and are fixing (it has one.cpp and two.cpp files instead of
test.cpp, so we need to add a custom run script).

We can fix this by just using a struct template with static constexpr
data members.

A couple of other changes here: Use is_floating_point_v instead
of ::value, and _Float_abs_bits doesn't need to static_cast at the end
(the inputs to bitwise-AND are the same type and are unsigned int or
unsigned long long, so there's no integral promotions involved).

cmath: Add Oxford comma to comment.

numeric: Now that we require is_integral_v, _Arithmetic should be
renamed to _Integral.