/std:c++latest configs are /permissive or /permissive-
So our test coverage won't be damaged when the compiler changes `/std:c++latest` to imply `/permissive-`. I do so by adding `/permissive-` to every `/std:c++latest` command line that doesn't yet have it, and then reverting one configuration in each matrix to `/permissive`. It seems sensible to me that the majority of our coverage should change from permissive to strict mode to reflect the change in compiler default.