Adjust backticks, drop ARM/ARM64 instructions.
According to https://github.github.com/gfm/#code-spans , we don't need
4 backticks; 2 will work. We can also get the space trimmed away - we
need a single space before and after for that to work. Finally,
capitalize Ctrl.

Drop the mention of adding `boost-math:arm-windows boost-math:arm64-windows`
to the `vcpkg.exe install` command - this requires ARM/ARM64 support to
be installed in VS, it has to be performed as a separate
`vcpkg.exe install` command (because it requires `boost-build:x86-windows`,
installed by the first command), this isn't mentioned in the
command-line instructions, and it is unnecessary for most users.

We can also drop "Assuming you are targeting x86 and x64," so this
exactly matches the command-line phrasing.