generator refactoring
Restore kernel mode operation by defining `unhandled_exception` only when `_KERNEL_MODE` is not defined. (When `unhandled_exception` is defined, the compiler will wrap the coroutine body in a `try...catch` as depicted in [dcl.fct.def.coroutine]/5, which it then diagnoses as ill-formed when unwinding is disabled as in `/kernel`.)

Fixes #1323.

Some cleanup while we're here:
* Use STL style:
  * Prefer not to deduce simple return types
  * `const` west
  * introduce template type parameters with `class`
  * Annotate functions that don't throw exceptions `noexcept`
  * remove excess empy lines
  * Use `_Capital_snake` names for non-template-parameters
  * `generator::promise_type` only has members `_Exception` and `_Rethrow_if_exception` to "delay" exception delivery so as to avoid triggering VSO-1172852. They aren't needed in kernel mode.
* `generator::iterator` should have _explicit_ conversion from `coroutine_handle` rather than implicit for good design hygiene. The conversion from `nullptr_t` is unnecessary.

Note that the story for `/EHs-` without `/kernel` is less clear and currently not supported.