In _Move_assign overloads, restore comparator copy assignment.
* It's okay for container assignment to require comparator assignment.

* Destroy-and-reconstruct is exception-unsafe - if reconstructing the
comparator throws an exception (e.g. because the comparator attempts to
allocate memory), then the container will be left with a destroyed
comparator, violating its invariants.

* It's intentional that container move construction/assignment performs
comparator copy construction/assignment, respectively. The Standard
(annoyingly!) doesn't talk about this clearly, but it's necessary to
preserve container invariants. The issue is that moved-from containers
still need to be in a valid-but-unspecified state (i.e. preconditionless
member functions can be called), so they can be clear()ed (which is
preconditionless) followed by inserting elements. This requires their
comparators to still be usable. However, a moved-from comparator may not
be usable - in particular, moved-from std::functions may be empty (and
are, in our implementation). Therefore, because comparators must always
be usable by the associative containers, they can't be moved-from.

This was unclear from the way I wrote up GH-1037; there I said
"`_Move_assign` also copy-assigns the comparator, then calls `_Swap_val`
anyways." and I should have said that the former was intentional but the
latter was undesirable.