stl/msbuild: Drop special_math.cpp comment.
This comment is increasingly confusing - it was added when the
BuildFiles below was moved to the top, which is the important part.
special_math.cpp doesn't need to be literally first.

We aren't going to arbitrarily reorder these lists, and in the long
term, our CMake build system will replace our legacy msbuild system.