Generalize the 80-bit long double in <complex> for Clang
The previous check works for the EDG-based Intel compiler ONLY, but
Clang and GCC both have the -mlong-double-80 options (see
https://godbolt.org/z/h7s4Te). This patch generalizes the 80 bit check
to work with any compiler that defines the __LDBL_DIG__ macro.

There are a couple of other macro options that are equally as opaque
(such as __LDBL_MANT_DIG__ and __LDBL_DECIMAL_DIG__), but this seemed
the most general. Note that it ALSO works in ICC, as well as all Clang
derivatives (such as Intel's DPCPP).