Downlevel standard coroutine support
Allow `<coroutine>` to be included if `_DOWNLEVEL_COROUTINES_SUPPORTED`
is defined even if `__cpp_lib_coroutine` is not. Allows use of standard
coroutines in C++14 and C++17 modes with a compiler that permits it.
Comparison operators are explicitly defined instead of <=> in these
modes.

Corresponding compiler changes are expected to come in 16.10.