<format>: Add const to some parameters
I added `const` to the some parameters where it made logical sense.
Sometimes the output iterator could be made constant, but that doesn't
seem correct to me. Also, I avoided making pointers to cstrings const,
because logically they need to be iterated over (as opposed to void*,
which doesn't). It feels a little verbose.