<format>: Increase the buffer size for specful floats (#1766)
Special thanks to @STL for helping define the limits of floating point
numbers.

The issue is basically that the buffer for specful floats was too small.
You can specify an arbitrary precision for some formatting types, which
means a lot more zeroes basically everywhere. If we only consider the
most precision needed to represent a floating point number full, we need
to consider the smallest double. It has an exponent of 2^-1074. As
proven in the comment, each power of 2 adds an additional decimal digit,
giving us a limit of 1074 decimal digits necessary to fully represent
the smallest number. On top of this, we need the decimal point, the
preceding 0, and a negative sign, bumping us to a total of 1077
characters. Any additional precision would only add 0s to the end of
this number. Now consider the largest number. It is much easier to
represent, clocking in at just 309 digits, none of which are decimal
points. However, if we format it as a fixed point with 1074 precision,
we would get 1074 zeroes at the end. To not worry about inspecting the
value we format, I have chosen a large enough buffer to accomodate the
largest number with the largest meaningful precision. Any precision past
this is just appending 0s.

Since only 'f' and '#g' care about the extra zeroes, the code change is
pretty managable.