<format>: Add localized specifier (#1767)
* <format>: Add localized specifier

Adds the 'L' specifier to specful writing. This works by precounting how
many "thousands" separaters are going to be necessary to write the whole
part of the number being formatted. Then the number is written piecewise
according to the grouping of the locale. This is the same for floating
points and integers so the implementation is reused.

The locale's grouping is a string(!) that specifies the number of digits
per grouping. Each character is used once, except the last one which is
reused as necessary. For example, in "en-US" the grouping is simply
`\3`, as numbers are divided by 3s. In "hi-IN" the groupings are `\3\2`.
This means that the hundreds are one group, then 2 digits are grouped
after that (e.g. 10,20,300).

Test are currently failing because the locale object exists in a
different DLL than the headers being used (when the test matrix mixes up
the `_ITERATOR_DEBUG_LEVEL`). This corrupts the stack and leads to bad
things.

The code is getting a little messy and probably overly complicated. Once
all the formatting specific stuff is put it, it would be a good idea to
go back and clean up the code. I don't want to do any major refactors
currently with multiple PRs in flight and being feature incomplete, but
once that's done I can look at it with fresh eyes.

* Make tests work

* Comments

* Update tests/std/tests/P0645R10_text_formatting_formatting/test.cpp

Remove extraneous `std::`-qualification

Co-authored-by: Casey Carter <cartec69@gmail.com>