<format>: Fix arg overload resolution (#1804)
* <format>: Fix arg overload resolution

There seems to be a compiler bug picking the wrong function in an
overload set. I've filed it here: . This sidesteps the bug, and also
fixes a latent bug with `_Get_format_arg_storage_typee`.

There is also a fix for the packed index, that wasn't very packed
because the types of the bitset were different. This should save space
as intended.

* Comments