<format>: Allow move-only output iterators
Now we move around the iterator we get instead of copying it. I had to move a few functions from their home in algorithm and xmemory into xutility. I hope I got the comments/guards correct, but please double check those.  The testing is done by making sure we don't use-after-move a custom iterator and that all the printing functions are instantiated.