Work around VSO-1307828 "Standard Library Header Units: ICE with queue's friend bool operator==".
The workaround is to replace queue/stack's unusual scheme of granting
friendship to function specializations, with a public (but _Ugly) accessor.

Retaining the original code for Clang/EDG alongside the workaround code
for C1XX would be a headache, so I've simply marked this as TRANSITION;
we can restore the original code from git history when this bug is fixed.