<chronat>: Add year and year_month_day formatting
This one is a biggie. The main things changed are the way that
specifiers are handled and delegated.

The general idea behind formatting time is that you can take segments of
a type and format them individually. For example, you can take the year
out of year_month_day and do the exact same operations you can do with a
normal year. The `_Is_type_valid` function recursively checks if a
parent type can be formatted by its children. The other functions didn't
really need much more finessing, the `tm` structure already has all the
fields we need to hold all the time info (simultaniously) and the
formatters work off that.

The big change here is in moving some "basic" formatters into our own
function and not relying on `get_time` to format them. The main reason
is that `get_time` does not play with invalid ranges, at all. A day of
`40` is always illegal, but we need to be able to format it, especially
in the face of `operator <<`. We could have kept what we had before, but
then it becomes a clear problem that we cannot use `%F` for a
`year_month_day` that has an invalid day, so I am seperating all the
integral formatters out into that function. Again, because of the nested
nature of times, we recurse in this function.

Note that function currently uses `format_to` in probably a very
inneficient way. I am all ears on how to improve that.