chronat: Clock formatting
Adds formatting for clocks!

Moved _Custom_write into the formatter so it can do special things
(write timezones).

_Write_seconds now writes leap seconds as 60 for utc clock time points.

Taught _Fill_tm to work with time_points (in reality only system_clock
and local_clock work).

Add operator<< for all clocks except local-time-format-t (because I'm
still not sure what that is).

The base formatter stores a timezone abbreviation. This is only useful
for the clocks, but it seemed like the simplest way to implement this
feature.