Add hh_mm_ss formatter
Adds the formatter with special bounds check so that `put_time` doesn't
assert.

Small bugfix/simplification for when `ok()` needs to be checked. Before
we call `put_time`, we should always bounds check.

Custom %S writer added (and %T calls it too). This is necessary for
proper fractional second printing.