chronat: Allow day from month_day_last conditionally
Thanks to matt and stat on discord for mentioning these edge cases.

I originally thought that you can't get a day from `month_day_last`, but
it clearly makes sense, for most months. I changed some machinery around
so that we can always intercept a specifier, which means that we need to
do manual checking in the writer for localization.

This PR is more to illustrate how a specifier may be intercepted, and
that testing need only be done for specific types.