chronat: Add duration formatter
Special specifiers for duration are `j q Q`.

Otherwise duration is very similar to `hh_mm_ss` except that times are
interpreted as time from midnight. I thought this would mean that
negative times are yesterday, but we just append a `-` instead, which
means we should round instead of flooring to a day when computing
hh/mm/ss.

Other behavior is pretty simple.