<chrono>: Add %j support and expand %aAuw
Adds %j support to all the classes that can benefit from it.
This means all year_month_day* but and month_last (but only for
January).

Changes %aAuw to throw on invalid weekdays, but still work if the
weekday is explicitly ok (so a bad year_month_day will always throw but
a bad year_month_weekday can still print the weekday if that portion is
ok).