<chrono>: Fix utc_clock seconds formatting
Seconds were being formatted incorrectly because we incorrectly assumed
that a day is 24*60*60 seconds (which is what flooring does) when in
reality they aren't always (like if there was a leap second). This error
compounds over time as leap seconds are inserted, so we need to add the
elapsed leap seconds at any given time point to get the correct seconds.

This issue doesn't affect the `tm` structure because it is constructed
out of a system time (`utc_clock::to_sys`), which has 24*60*60 second
days.