Return _Nontrivial_dummy_type to <optional>
We pushed this type up into `<xutility>` to use implementing the `optional`-alikes in `<ranges>`. After realizing that the updated C++20 `constexpr` rules mean we can avoid activating any union alternative in a `constexpr` constructor if we like, we removed the uses in `<ranges>`. This PR simply moves the type back to `<optional>` since it's no longer used elsewhere.