Fix Efficient access to streambuf buffer
Currently we were always taking the buffer from the moved in string, but would only take it when the buffer could be read or written to.

Consequently if that was not the case we would leak the allocated memory.

Also simplify the code a bit, we know that the capacity `_Res` will always be nonzero due to the was `_Release_to_buffer` works

Also add tests for the error cases