Implement compile-time restraints for reference types in containers and allocator
Fixes #2408.

This PR implements a message (or more, depending on structure)  when you try to instantiate an object which has inherent contradicting type requirements (ex. "pointer to reference").

I tried not to be pedantic with the implementation, but I might've went wrong with that, thus I'm sorry if I did.

Notes: I have noticed different order for implementing static asserts in containers, sometimes it's before required types and right after, if there is, friend classes, and sometimes, it's after required types and in different access modifiers (not like it changes anything, just a nitpick).