Update comments and test coverage for `/await:strict`.
Thanks to @cpplearner for noticing this. Coroutines are usually C++20
but are also activated by `/await:strict`, so we permanently need to
check `__cpp_impl_coroutine` - this is not a temporary thing until
Clang 14 is released. (Tests that are C++20-specific can be simplified
after Clang 14, so they still have TRANSITION comments.) Finally, we
should also add a new configuration to
`VSO_0157762_feature_test_macros`.