<source_location>: Support clang 15.0.0 with GCC like __builtin_source_location() builtin.
clang implements GCC's __builtin_source_location, the assumed implementation for clang is not the same as source_location in MSVC STL. (Probably for libcxx and libstdc++) To avoid ABI breaking issues, we have to do something instead of just using the supposed implementation in GNU C++ library's source_location.

However, it is still possible to get around of issue for supporting clang with __builtin_source_location() by hacking the compiler around.