Cite GitHub issues with GH-nnn.
This is helpful with the GitHub Pull Requests And Issues extension for VS Code.

In GH_001001_random_rejection_rounding/test.cpp, this adds a comment.

In GH_000431_equal_family/test.cpp, this renames test_gh1523() to test_gh_1523() for consistency.

In Dev09_126254_persistent_aux_allocators/test.cpp, this omits a mention of a GitHub username.