[copy_file] allow FILE_SHARE_*
The existing code does not allow other programs to hold read permissions
on _Source or _Target when
copy_options = overwrite_existing or skip_existing.
We could very easily allow reading, and could also allow writing for
skip_existing, except for a clever implementation of equivalent,
so just switch to a dumb implementation of equivalent here.