Call atexit() within `if (!initialize_attempted)`.
lock_and_uninitialize() is highly robust to partial initialization, so we can register it immediately.

Note that if atexit() fails, we haven't initialized anything, so we don't need to call uninitialize() to clean up. Instead, we can immediately return false.

Finally, note that if atexit() fails, we leave initialize_attempted set to true (uninitialize() would unset it), so we will never attempt initialization again.